Move ProjectTopbar fragments to own directory
