+++ app/src/DesignSystem/Components/ProjectTopbar/MenuItem.svelte
